Добавили функции в модуле map, добавили модуль page-state
