+++ js/main.js
@@ -1,2 +1,4 @@
 import {showMap} from './map.js'
+import {deactivatePage} from './page-state.js'
+deactivatePage()
 showMap()
