+++ js/page-state.js
@@ -1,0 +1,38 @@
+import {TOKIO_CENTER} from './map.js'
+
+const adForm = document.querySelector('.ad-form');
+const mapFilters = document.querySelector('.map__filters');
+const filterSelect = mapFilters.querySelectorAll('select'); 
+const fieldsetBlock =  adForm.querySelectorAll('fieldset');
+const addressBlock = document.querySelector('#address');
+
+//Переводим страницу в неактивное состояние
+const deactivatePage = () => {
+  filterSelect.forEach((select) => {
+    select.setAttribute('disabled', 'disabled')
+  })
+  mapFilters.querySelector('fieldset').setAttribute('disabled', 'disabled')
+  mapFilters.classList.add('map__filters--disabled')
+  fieldsetBlock.forEach((block) => {
+    block.setAttribute('disabled', 'disabled')
+  })
+  adForm.classList.add('ad-form--disabled')
+}
+
+// Переводим страницу в активное состояние
+const activatePage = (status) => {
+  if (!status) {
+    adForm.classList.remove('ad-form--disabled')   
+    fieldsetBlock.forEach((block) => {
+      block.removeAttribute('disabled', 'disabled')
+      mapFilters.querySelector('fieldset').removeAttribute('disabled', 'disabled')
+      mapFilters.classList.remove('map__filters--disabled')
+      filterSelect.forEach((select) => {
+        select.removeAttribute('disabled', 'disabled')
+      })
+    })
+    addressBlock.value = [TOKIO_CENTER.x.toFixed(5), TOKIO_CENTER.y.toFixed(5)]
+  }
+}
+
+export {activatePage, deactivatePage, addressBlock}
